Overhaul initialization logic
As of scyjava 1.1.0, we can defer initialization of Java-dependent
data structures until after the JVM is available.

This means we no longer need to worry about which modules and submodules
we import when; just import what you want at the top, as usual for
Python, and everything will "just work". The only exception is if you
try to access a Java-dependent data structure before calling any scyjava
or imglyb functions -- in that case, you can call scyjava.start_jvm()
explicitly beforehand to ensure everything is set up.