Wrap RAI as numpy array
We want to be able to link a RAI in java to a Numpy array in Python.
I saw three potential solutions to accomplishing this.

1. We let the RAI and the numpy array point to the same memory location.
* This is only possible if we can know the memory location, and this is
actually what ImgLibReferenceGuard does. But we don't know the memory
location for most RAIs. (Of course, if we do, we will use this instead)

2. We convert the RAI to (nested) lists, and pass that into np.array()
  or some similar function.
* This will make a copy of those (nested) lists, which does not link the
structures for us.

Thus, I pursue the third solution:

3. Wrap the RAI as a class implementing np.ndarray.